Move internal/core/analysis under internal/core/analysis_api
